09._Environment/02._python/pyproject.toml
[project]
name = "02-python"
version = "0.1.0"
description = "Working with environment variable in Python"
authors = [
    {name = "Samim S.",email = "samimsaidi2660@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "02"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
